Maze Generation using HK and Seed
This code contain a custom implementation of random number generation
using seed value so that it will generate same maze for same seed
value.
Seed is the one time value give to the random number generator
and it will assig the new seed value for every call so that the
sequence of random number generated while calling the function is same
everytime for a given seed value.